camera fixed
